Add support for the basic watchdog which runs one server at a time
