+++ src/Tgstation.Server.Host/appsettings.json
@@ -4,7 +4,8 @@
     "GitHubAccessToken": null,
     "SetupWizardMode": "AutoDetect",
     "ByondTopicTimeout": 5000,
-    "RestartTimeout": 10000
+    "RestartTimeout": 10000,
+    "UseExperimentalWatchdog" : false
   },
   "FileLogging": {
     "Directory": null, //use the default path
